move send_close_signal to execute
